+++ src/main/java/org/yjhking/pethome/org/mapper/EmployeeMapper.java
@@ -17,7 +17,7 @@ public interface EmployeeMapper {
      * @param id 主键
      * @return 删除的id
      */
-    int deleteByPrimaryKey(Long id);
+    Long deleteByPrimaryKey(Long id);
     
     /**
      * 新增
@@ -25,7 +25,7 @@ public interface EmployeeMapper {
      * @param employee 新增的数据
      * @return 新增的id
      */
-    int insertSelective(Employee employee);
+    Long insertSelective(Employee employee);
     
     /**
      * 通过主键查询
@@ -41,7 +41,7 @@ public interface EmployeeMapper {
      * @param employee 更新的数据
      * @return 更新的id
      */
-    int updateByPrimaryKeySelective(Employee employee);
+    Long updateByPrimaryKeySelective(Employee employee);
     
     /**
      * 查询全部
