+++ src/main/java/org/yjhking/pethome/org/service/DepartmentService.java
@@ -18,7 +18,7 @@ public interface DepartmentService {
      * @param id 主键
      * @return 删除的id
      */
-    int deleteByPrimaryKey(Long id);
+    Long deleteByPrimaryKey(Long id);
     
     /**
      * 增加数据
@@ -26,7 +26,7 @@ public interface DepartmentService {
      * @param department 部门
      * @return 增加的id
      */
-    int insertSelective(Department department);
+    Long insertSelective(Department department);
     
     /**
      * 通过主键查询
@@ -42,7 +42,7 @@ public interface DepartmentService {
      * @param department 更新的数据
      * @return 更新的id
      */
-    int updateByPrimaryKeySelective(Department department);
+    Long updateByPrimaryKeySelective(Department department);
     
     /**
      * 查询全部
